feat(DialogGuidance): Moved code for student-friendly idea text into separate component
