+++ src/assets/wise5/components/common/cRater/edit-crater-idea-descriptions/edit-crater-idea-descriptions.component.ts
@@ -1,22 +1,75 @@
-// import { Component, Input, OnInit } from '@angular/core';
-// import { CdkDrag, CdkDragDrop, CdkDropList, moveItemInArray } from '@angular/cdk/drag-drop';
-// import { Subject, Subscription } from 'rxjs';
-// import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
-// import { TeacherProjectService } from '../../../../services/teacherProjectService';
-// import { MatDialog } from '@angular/material/dialog';
-// import { MatIconModule } from '@angular/material/icon';
-// import { generateRandomKey } from '../../../../common/string/string';
-// import { CRaterIdea } from '../CRaterIdea';
+import { CdkTextareaAutosize } from '@angular/cdk/text-field';
+import { Component, Input, OnInit } from '@angular/core';
+import { ComponentContent } from '../../../../common/ComponentContent';
+import { CRaterIdea } from '../CRaterIdea';
+import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatCardModule } from '@angular/material/card';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { FormsModule } from '@angular/forms';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { Subject, Subscription } from 'rxjs';
+import { TeacherProjectService } from '../../../../services/teacherProjectService';
 
-// @Component({
-//   selector: 'edit-crater-idea-descriptions',
-//   standalone: true,
-//   imports: [MatIconModule],
-//   templateUrl: './edit-crater-idea-descriptions.component.html',
-//   styleUrl: './edit-crater-idea-descriptions.component.scss'
-// })
-// export class EditCRaterIdeaDescriptionsComponent {
-//   @Input() ideaDescriptions: CRaterIdea[] = [];
+@Component({
+  imports: [
+    CdkTextareaAutosize,
+    FlexLayoutModule,
+    FormsModule,
+    MatCardModule,
+    MatInputModule,
+    MatFormFieldModule,
+    MatIconModule,
+    MatTooltipModule
+  ],
+  selector: 'edit-crater-idea-descriptions',
+  standalone: true,
+  templateUrl: './edit-crater-idea-descriptions.component.html',
+  styleUrl: './edit-crater-idea-descriptions.component.scss'
+})
+export class EditCRaterIdeaDescriptionsComponent implements OnInit {
+  @Input() componentContent: ComponentContent;
+  @Input() ideaDescriptions: CRaterIdea[] = [];
+  inputChanged: Subject<string> = new Subject<string>();
+  subscriptions: Subscription = new Subscription();
 
-//   protected createNewDescription(): void {}
-// }
+  constructor(protected teacherProjectService: TeacherProjectService) {}
+
+  ngOnInit(): void {
+    this.subscriptions.add(
+      this.inputChanged.pipe(debounceTime(1000), distinctUntilChanged()).subscribe(() => {
+        this.teacherProjectService.nodeChanged();
+      })
+    );
+    // this.i;
+  }
+
+  ngOnDestroy(): void {
+    this.subscriptions.unsubscribe();
+  }
+
+  protected addNewIdeaDescription(): void {
+    const newIdeaDescription = this.createNewIdea();
+    this.ideaDescriptions.splice(this.getIdeaIndex(), 0, newIdeaDescription);
+    this.teacherProjectService.nodeChanged();
+  }
+
+  private createNewIdea(): CRaterIdea {
+    const idea = new CRaterIdea('', null);
+    idea.description = '';
+    return idea;
+  }
+
+  protected getIdeaIndex(): number {
+    return this.ideaDescriptions.length;
+  }
+
+  deleteIdeaDescription(ideaIndex: number): void {
+    if (confirm($localize`Are you sure you want to delete this feedback rule?`)) {
+      this.ideaDescriptions.splice(ideaIndex, 1);
+      this.teacherProjectService.nodeChanged();
+    }
+  }
+}
