+++ src/assets/wise5/components/dialogGuidance/edit-dialog-guidance-advanced/edit-dialog-guidance-advanced.component.html
@@ -1,78 +1,3 @@
 <edit-component-constraints [componentContent]="component.content" />
 <edit-component-json [component]="component"></edit-component-json>
-
-<div class="idea-descriptions notice-bg-bg">
-  <h5 fxLayout="row" fxLayoutAlign="start center" fxLayoutGap="4px">
-    <span i18n>Idea Descriptions</span>
-    <button
-      mat-icon-button
-      color="primary"
-      matTooltip="Add a new idea description"
-      i18n-matTooltip
-      matTooltipPosition="after"
-      (click)="addNewIdeaDescription()"
-    >
-      <mat-icon>add_circle</mat-icon>
-    </button>
-    <span fxFlex></span>
-  </h5>
-  <ul>
-    @for (
-      idea of ideaDescriptions;
-      track $index;
-      let ideaIndex = $index, first = $first, last = $last
-    ) {
-      <li class="idea">
-        <mat-card appearance="outlined" class="idea-content" fxLayout="row wrap" fxLayoutGap="8px">
-          <div
-            class="text-secondary"
-            fxLayout="column"
-            fxLayoutAlign="start center"
-            fxLayoutGap="22px"
-          >
-            <span class="mat-subtitle-1">{{ ideaIndex + 1 }}</span>
-          </div>
-          <div fxLayout="column" fxLayoutAlign="center start" fxLayoutGap="8px" fxFlex>
-            <mat-form-field class="idea-input form-field-no-hint" appearance="fill">
-              <mat-label i18n>Idea #</mat-label>
-              <input matInput [(ngModel)]="idea.name" (ngModelChange)="inputChanged.next($event)" />
-            </mat-form-field>
-            <mat-form-field class="idea-input form-field-no-hint" appearance="fill">
-              <mat-label i18n>Student-Friendly Description</mat-label>
-              <textarea
-                matInput
-                [(ngModel)]="idea.description"
-                (ngModelChange)="inputChanged.next($event)"
-                cdkTextareaAutosize
-              >
-              </textarea>
-            </mat-form-field>
-          </div>
-          <div fxLayout="column" fxLayoutAlign="start center">
-            <button
-              mat-icon-button
-              i18n-matTooltip
-              matTooltip="Delete idea description"
-              matTooltipPosition="before"
-              (click)="deleteIdeaDescription(ideaIndex)"
-            >
-              <mat-icon>clear</mat-icon>
-            </button>
-          </div>
-        </mat-card>
-      </li>
-    }
-  </ul>
-  <div [hidden]="getIdeaIndex() === 0">
-    <button
-      mat-icon-button
-      color="primary"
-      matTooltip="Add a new idea description"
-      i18n-matTooltip
-      matTooltipPosition="above"
-      (click)="addNewIdeaDescription()"
-    >
-      <mat-icon>add_circle</mat-icon>
-    </button>
-  </div>
-</div>
+<edit-crater-idea-descriptions [ideaDescriptions]="ideaDescriptions" />
